+++ sequential-parser/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Revision history for sequential-parser
 
-## 0.1.0.0 -- YYYY-mm-dd
+## 0.1.0.0 -- 2023-02-16
 
-* First version. Released on an unsuspecting world.
+* First implementation of a table-driven LL(1) parser that will just check if the string can be
+  parsed and give the rules used.
+